build: split workflows into multiple files
Splits up the workflow in two parts:
- one workflow for testing and building (build-and-test.yaml)
- one for publishing the documentation (publish-documentation.yaml)
- (there will be a 3rd one for releasing) (release.yaml)

What we need to achieve is running jobs conditional.
While github allows checking for branch name there is no way to trigger a job manually. But we need this to release.
Workflows can be triggered manually and they can be reused.

We have 3 triggers we care about that should to lead slightly different output:
- PR: just run checks
- Push to main: run checks and deploy pages
- Manual release trigger: run checks and publish npm packages

So the publish-documentation and release workflow are reusing the build-and-test workflow.