refactor(tabs-next): enforce using methods to select and deselect tabs
Previously, the active flag of all tabs implementation must have been a WritableSignal.
This was causing problems in the link tabs, as there everything is managed by the router.
So we needed an effect to react to changes on the active signal.

Wit this change, we now have methods to programmatically select and deselect tabs.
Those methods never cause the emitting of events and could also be used by applications.
The active flag is now also readonly in the base class to prevent setting it directly
if the router is used.